Move common, useful parts into a shared file
