+++ src/hivpy/population.py
@@ -4,7 +4,8 @@ from typing import Callable, Dict
 
 import pandas as pd
 
-from .demographics import DemographicsModule, SexType
+from .common import SexType
+from .demographics import DemographicsModule
 from .hiv_status import HIVStatusModule
 from .sexual_behaviour import SexualBehaviourModule
